feat: add one layout for all pages
