+++ app/asteroid/[id]/page.module.css
@@ -1,5 +1,5 @@
 .title__description {
-  font-family: var(--font-helvetica);
+
   font-size: 14px;
   font-style: normal;
   line-height: 24px;
@@ -7,12 +7,12 @@
 
 .list__head {
   margin: 10px 0;
-  font-family: var(--font-helvetica);
+
 }
 
 .list__item__header {
   color: var(--White-Base, #FFF);
-  font-family: var(--font-helvetica);
+
   font-size: 24px;
   font-style: normal;
   font-weight: 700;
