+++ app/components/header/Header.module.css
@@ -11,7 +11,7 @@
 
 .header__title {
   color: var(--Orange-Base);
-  font-family: 'Passion One', 'Courier New';
+  font-family: 'Passion One', 'Arial Condensed', 'Onyx', 'Arial';
   font-size: 32px;
   font-style: normal;
   font-weight: 400;
